+++ h2integrate/core/h2integrate_model.py
@@ -1,3 +1,4 @@
+import importlib.util
 from pathlib import Path
 
 import yaml
@@ -23,6 +24,9 @@ class H2IntegrateModel:
     def __init__(self, config_file):
         # read in config file; it's a yaml dict that looks like this:
         self.load_config(config_file)
+
+        # load custom models
+        self.collect_custom_models()
 
         # create site-level model
         # this is an OpenMDAO group that contains all the site information
@@ -59,10 +63,50 @@ class H2IntegrateModel:
 
         # Load each config file as yaml and save as dict on this object
         self.driver_config = load_driver_yaml(config_path.parent / config.get("driver_config"))
-        self.technology_config = load_tech_yaml(
-            config_path.parent / config.get("technology_config")
-        )
+        self.tech_config_path = config_path.parent / config.get("technology_config")
+        self.technology_config = load_tech_yaml(self.tech_config_path)
         self.plant_config = load_plant_yaml(config_path.parent / config.get("plant_config"))
+
+    def collect_custom_models(self):
+        """
+        Collect custom models from the technology configuration.
+
+        This method loads custom models from the specified directory and adds them to the
+        supported models dictionary.
+        """
+
+        self.supported_models = supported_models.copy()
+
+        for tech_name, tech_config in self.technology_config["technologies"].items():
+            for model_type in ["performance_model", "cost_model", "financial_model"]:
+                if model_type in tech_config:
+                    model_name = tech_config[model_type].get("model")
+                    if model_name not in self.supported_models and model_name is not None:
+                        model_class_name = tech_config[model_type].get("model_class_name")
+                        model_location = tech_config[model_type].get("model_location")
+
+                        if not model_class_name or not model_location:
+                            raise ValueError(
+                                f"Custom {model_type} for {tech_name} must specify "
+                                "'model_class_name' and 'model_location'."
+                            )
+
+                        # Resolve the full path of the model location
+                        model_path = self.tech_config_path.parent / model_location
+
+                        if not model_path.exists():
+                            raise FileNotFoundError(
+                                f"Custom model location {model_path} does not exist."
+                            )
+
+                        # Dynamically import the custom model class
+                        spec = importlib.util.spec_from_file_location(model_class_name, model_path)
+                        module = importlib.util.module_from_spec(spec)
+                        spec.loader.exec_module(module)
+                        custom_model_class = getattr(module, model_class_name)
+
+                        # Add the custom model to the supported models dictionary
+                        self.supported_models[model_name] = custom_model_class
 
     def create_site_model(self):
         # Create a site-level component
@@ -120,20 +164,20 @@ class H2IntegrateModel:
                 tech_group = self.plant.add_subsystem(tech_name, om.Group())
                 self.tech_names.append(tech_name)
 
-                # Special HOPP handling for short-term
+                # Special handling for short-term components that share performance and cost models
                 if tech_name in combined_performance_and_cost_model_technologies:
-                    hopp_comp = supported_models[tech_name](
+                    comp = self.supported_models[tech_name](
                         plant_config=self.plant_config, tech_config=individual_tech_config
                     )
-                    tech_group.add_subsystem(tech_name, hopp_comp, promotes=["*"])
-                    self.performance_models.append(hopp_comp)
-                    self.cost_models.append(hopp_comp)
-                    self.financial_models.append(hopp_comp)
+                    tech_group.add_subsystem(tech_name, comp, promotes=["*"])
+                    self.performance_models.append(comp)
+                    self.cost_models.append(comp)
+                    self.financial_models.append(comp)
                     continue
 
                 # Process the technology models
                 performance_name = individual_tech_config["performance_model"]["model"]
-                performance_object = supported_models[performance_name]
+                performance_object = self.supported_models[performance_name]
                 tech_group.add_subsystem(
                     performance_name,
                     performance_object(
@@ -146,7 +190,7 @@ class H2IntegrateModel:
                 # Process the cost models
                 if "cost_model" in individual_tech_config:
                     cost_name = individual_tech_config["cost_model"]["model"]
-                    cost_object = supported_models[cost_name]
+                    cost_object = self.supported_models[cost_name]
                     tech_group.add_subsystem(
                         cost_name,
                         cost_object(
@@ -161,7 +205,7 @@ class H2IntegrateModel:
                     if "model" in individual_tech_config["financial_model"]:
                         financial_name = individual_tech_config["financial_model"]["model"]
 
-                        financial_object = supported_models[financial_name]
+                        financial_object = self.supported_models[financial_name]
                         tech_group.add_subsystem(
                             f"{tech_name}_financial",
                             financial_object(
@@ -263,7 +307,7 @@ class H2IntegrateModel:
                 connection_name = f"{source_tech}_to_{dest_tech}_{transport_type}"
 
                 # Create the transport object
-                connection_component = supported_models[transport_type]()
+                connection_component = self.supported_models[transport_type]()
 
                 # Add the connection component to the model
                 self.plant.add_subsystem(connection_name, connection_component)